requirements.txt
flask
diffusers
transformers
torch
accelerate
pillow